Sort inspector field and split HDCameraUI
