+++ com.unity.render-pipelines.high-definition/HDRP/Editor/Camera/HDCameraUI.Drawers.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1330afbc510383b4b9e477fe8b582100
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
