+++ com.unity.render-pipelines.high-definition/HDRP/Editor/Camera/HDCameraUI.Skin.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 97b4d7e55cbca7b4cb8bbc4d28c036f2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
